// src/types/api/user.ts
/**
 * LifeMapWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface User { 
    UserID?: number;
    Username?: string;
    UserTypeID?: string;
    EmailAddress?: string;
    FirstName?: string;
    LastName?: string;
    StreetAddress?: string;
    City?: string;
    StateRegion?: string;
    Country?: string;
    PostalCode?: string;
    PhoneNumberMobile?: string;
    PhoneNumberHome?: string;
    PhoneNumberWork?: string;
    Company?: string;
    AvatarURL?: string;
    ProfilePhoto?: string;
    AddressListID?: number;
    PhoneNumberListID?: number;
    CommentListID?: number;
    CreationDate?: Date;
    CreatedBy?: number;
    LastModifiedDate?: Date;
    LastModifiedBy?: number;
    CurrentPassword?: string;
    NewPassword?: string;
    ConfirmPassword?: string;
    PhotoURL?: string;
    LogoURL?: string;
    FirmID?: number;
    FirmName?: string;
    
}